Refactor afor Maven use
