--- Web_Designer/bin/.gitignore
@@ -1,1 +1,0 @@
-/tutoriel/
