--- Web_Designer/pom.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>Web_Designer</groupId>
-	<artifactId>Web_Designer</artifactId>
-	<version>0.0.1</version>
-	
-	<dependencies>
-		<dependency>
- 			<!-- jsoup HTML parser library @ https://jsoup.org/ -->
-  			<groupId>org.jsoup</groupId>
-			<artifactId>jsoup</artifactId>
-			<version>1.13.1</version>
-		</dependency>
-	</dependencies>
-	
-</project>
