+++ deps/Animation_example/largetux-walk-left-0.png
