+++ deps/Animation_example/largetux-walk-left-1.png
