+++ deps/Animation_example/largetux-walk-left-2.png
