+++ deps/Animation_example/largetux-walk-left-3.png
