+++ deps/Animation_example/largetux-walk-left-4.png
