+++ deps/Animation_example/largetux-walk-right-0.png
