+++ deps/Animation_example/largetux-walk-right-1.png
