+++ deps/Animation_example/largetux-walk-right-2.png
