+++ deps/Animation_example/largetux-walk-right-3.png
