+++ deps/Animation_example/largetux-walk-right-4.png
